+++ src/elements/data-structure.js
@@ -1,5 +1,5 @@
 export default function (namespace) {
-  class DataStructure extends namespace.BaseElement {
+  class DataStructure extends namespace.Element {
     constructor(...args) {
       super(...args);
       this.element = 'dataStructure';
@@ -7,34 +7,6 @@ export default function (namespace) {
       if (this.content !== undefined) {
         this.content = namespace.toElement(this.content);
       }
-    }
-
-    toValue() {
-      return this.content && this.content.toValue();
-    }
-
-    toRefract() {
-      const refract = super.toRefract();
-
-      if (Array.isArray(this.content)) {
-        refract.content = this.content.map(item => item.toRefract());
-      } else {
-        refract.content = this.content.toRefract();
-      }
-
-      return refract;
-    }
-
-    fromRefract(doc) {
-      super.fromRefract(doc);
-
-      if (Array.isArray(doc.content)) {
-        this.content = doc.content.map(item => namespace.fromRefract(item));
-      } else {
-        this.content = namespace.fromRefract(doc.content);
-      }
-
-      return this;
     }
   }