+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.4.0
+
+- Updated to support minim 0.17.1.
+
 # 0.3.1 - 2017-05-12
 
 - Added API metadata convenience function